pyproject.toml
[project]
name = "browser-use-fetcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "browser-use>=0.1.0",
    "python-dotenv>=1.1.0",
    "langchain-openai>=0.3.11",
]
